+++ kernel/watch_queue.c
@@ -126,7 +126,7 @@ static bool post_one_notification(struct watch_queue *wqueue,
 
 	buf = &pipe->bufs[head & mask];
 	buf->page = page;
-	buf->private = (unsigned long)wqueue;
+	buf->private = (uintptr_t)wqueue;
 	buf->ops = &watch_queue_pipe_buf_ops;
 	buf->offset = offset;
 	buf->len = len;
